misc: fix some compiler warnings
  libsmartcols/samples/fromfile.c:59:2: warning: passing argument 3 of 'string_to_bitmask' from incompatible pointer type
  text-utils/pg.c:79:0: warning: "TABSIZE" redefined
  libblkid/src/read.c:455:13: warning: 'debug_dump_dev' defined but not used [-Wunused-function]
  libblkid/src/probe.c:769:13: warning: unused function 'cdrom_size_correction' [-Wunused-function]
  /usr/include/sys/termios.h:3:2: warning: "this file includes <sys/termios.h> which is deprecated, use <termios.h> instead" [-W#warnings]

+++ libsmartcols/samples/fromfile.c
@@ -38,7 +38,7 @@ static const struct column_flag flags[] = {
 	{ "none",	0 }
 };
 
-static long name_to_flag(const char *name, long unsigned int namesz)
+static long name_to_flag(const char *name, size_t namesz)
 {
 	size_t i;
